Addition of some C++ files
cfp.cpp accesses the postgresql db via libpgxx.
procj.cpp uses libjpeg to access a jpeg file (in a way that could be from a postgres read, but atm is from a file) and does something to nxn squares of it, and writes it to a ppm file (for debug purposes). 
test.jpg ios a picture of my cat, Potassium, sitting on floor. Used by procj.cpp as a test file.